requirements.txt
asgiref==3.2.10
Django==3.0.7
djangorestframework==3.11.0
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1